fix: component code in correct file
